+++ src/Logging/EventsUdpLogstashLogger.php
@@ -7,7 +7,6 @@ namespace AvtoDev\EventsLogLaravel\Logging;
 use Exception;
 use Monolog\Logger;
 use InvalidArgumentException;
-use Monolog\Formatter\LogstashFormatter;
 use AvtoDev\EventsLogLaravel\Contracts\LoggerContract;
 use AvtoDev\EventsLogLaravel\Logging\Formatters\EventsLogstashFormatter;
 
@@ -29,9 +28,8 @@ class EventsUdpLogstashLogger implements LoggerContract
         $formatter = new EventsLogstashFormatter(
             $config['formatter']['app_name'] ?? $this->getAppName() ?? 'app',
             $config['formatter']['system_name'] ?? null,
-            $config['formatter']['extra_prefix'] ?? false,
-            $config['formatter']['context_prefix'] ?? null,
-            $config['formatter']['version'] ?? LogstashFormatter::V1
+            $config['formatter']['extra_prefix'] ?? 'extra',
+            $config['formatter']['context_prefix'] ?? 'context'
         );
 
         $handler = new Handlers\UdpHandler(