+++ src/Logging/Formatters/EventsLogstashFormatter.php
@@ -4,21 +4,30 @@ declare(strict_types = 1);
 
 namespace AvtoDev\EventsLogLaravel\Logging\Formatters;
 
-class EventsLogstashFormatter extends AbstractLogstashFormatter
+final class EventsLogstashFormatter extends AbstractLogstashFormatter
 {
     /**
-     * {@inheritdoc}
-     *
-     * @return array<string, mixed> Message
+     * Records entry type.
      */
-    protected function modifyParentMessage(array $parent_message, array $record): array
+    public const ENTRY_TYPE = 'event';
+
+    /**
+     * Formats a log record.
+     *
+     * @param array<string, mixed> $record
+     *
+     * @return string
+     */
+    public function format(array $record): string
     {
-        $parent_message['entry_type'] = 'event';
+        $formatted = $this->toLogstashFormat($record);
+
+        $formatted['entry_type'] = self::ENTRY_TYPE;
 
         if (isset($record['context']['event']) && \is_array($event_data = $record['context']['event'])) {
-            $parent_message[$this->extraPrefix . 'event'] = $event_data;
+            $formatted[$this->extraKey . 'event'] = $event_data;
         }
 
-        return $parent_message;
+        return $this->toJson($formatted) . "\n";
     }
 }